Updated Engine.cpp Removed SDL_Image.h To internal Folder Added imports For SFML Updated Makefile
Good morning everyone, today I’ve improved the engine and reworked part of the project. It’s all set and ready to go. SDL_Image.h is now supported in the includes, and I’ve added -lSDL2_image to the Makefile. The engine now also supports SFML imports, preparing for the 1.0 update.